backend/requirements.txt
Flask>=2.3,<3
gunicorn>=21.2,<22
psycopg2-binary>=2.9,<3
requests>=2.31,<3
